Moved files across
The code and files to make the SVG animation using the Animation library provided by the lecturer. The exported SVG file 'animation.svg' is the finished product. The rest is the code which would require the software 'WinGHCI' to compile. But we already have the output so it's fine.
I had made this repo in Gitlab but IDK how to move them over to here so I just copied the files in the folder to this one.